move genome parsing code to separate module
